move list configs to a separate hook
