+++ lib/FindLocation/FindLocationLookup.js
@@ -1,7 +1,7 @@
 import keyBy from 'lodash/keyBy';
 import PropTypes from 'prop-types';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import { FormattedMessage, useIntl } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
 import { ColumnManager } from '@folio/stripes/smart-components';
 
@@ -14,13 +14,14 @@ import {
   useLibraries,
 } from '../hooks';
 import {
-  COLUMN_MAPPING,
-  COLUMN_NAMES,
   NON_TOGGLEABLE_COLUMNS,
   SORTABLE_COLUMNS,
 } from './configs';
 import { FindLocationFilters } from './FindLocationFilters';
-import { useLocationsRecords } from './useLocationsRecords';
+import {
+  useLocationsList,
+  useLocationsRecords
+} from './hooks';
 
 export const FindLocationLookup = (props) => {
   const {
@@ -35,7 +36,6 @@ export const FindLocationLookup = (props) => {
     ...rest
   } = props;
 
-  const intl = useIntl();
   const eventEmitter = useEventEmitter();
 
   const [selectedLocations, setSelectedLocations] = useState({});
@@ -63,40 +63,25 @@ export const FindLocationLookup = (props) => {
     Subscribe on changes in selected records.
   */
   useEffect(() => {
-    const eventType = EVENT_EMMITER_EVENTS.FIND_RECORDS_SELECTED_CHANGED;
     const callback = ({ detail: assignment }) => setSelectedLocations(assignment);
 
-    eventEmitter.on(eventType, callback);
+    eventEmitter.on(EVENT_EMMITER_EVENTS.FIND_RECORDS_SELECTED_CHANGED, callback);
 
     return () => {
-      eventEmitter.off(eventType, callback);
+      eventEmitter.off(EVENT_EMMITER_EVENTS.FIND_RECORDS_SELECTED_CHANGED, callback);
     };
   }, [eventEmitter]);
 
-  const columnMapping = useMemo(() => ({
-    ...COLUMN_MAPPING,
-    ...(
-      isMultiSelect
-        ? { [COLUMN_NAMES.isAssigned]: <FormattedMessage id="stripes-acq-components.find-location.results.column.assignment" /> }
-        : {}
-    ),
-  }), [isMultiSelect]);
-
-  const resultsFormatter = useMemo(() => ({
-    [COLUMN_NAMES.name]: ({ name }) => name,
-    [COLUMN_NAMES.code]: ({ code }) => code,
-    [COLUMN_NAMES.institution]: ({ institutionId }) => institutionsMap[institutionId]?.name,
-    [COLUMN_NAMES.campus]: ({ campusId }) => campusesMap[campusId]?.name,
-    [COLUMN_NAMES.library]: ({ libraryId }) => librariesMap[libraryId]?.name,
-    [COLUMN_NAMES.isActive]: ({ isActive }) => intl.formatMessage({ id: `stripes-acq-components.find-location.results.column.status.${isActive ? 'active' : 'inactive'}` }),
-    [COLUMN_NAMES.isAssigned]: ({ id }) => intl.formatMessage({ id: `stripes-acq-components.filter.assignment.${selectedLocations[id] ? 'assigned' : 'unassigned'}` }),
-  }), [
+  const {
+    columnMapping,
+    resultsFormatter,
+  } = useLocationsList({
     campusesMap,
     institutionsMap,
-    intl,
+    isMultiSelect,
     librariesMap,
     selectedLocations,
-  ]);
+  });  
 
   const isLoading = (
     isLocationsLoading
