+++ lib/FindLocation/hooks/index.js
@@ -1,0 +1,2 @@
+export { useLocationsList } from './useLocationsList';
+export { useLocationsRecords } from './useLocationsRecords';
