+++ lib/FindLocation/hooks/useLocationsList/index.js
@@ -1,0 +1,1 @@
+export { useLocationsList } from './useLocationsList';
