+++ lib/FindLocation/hooks/useLocationsRecords/index.js
@@ -1,0 +1,1 @@
+export { useLocationsRecords } from './useLocationsRecords';
