+++ lib/FindLocation/hooks/useLocationsRecords/useLocationsRecords.js
@@ -2,13 +2,13 @@ import identity from 'lodash/identity';
 import omit from 'lodash/omit';
 import { useEffect, useState } from 'react';
 
-import { EVENT_EMMITER_EVENTS } from '../constants';
+import { EVENT_EMMITER_EVENTS } from '../../../constants';
 import {
   useEventEmitter,
   useLocations,
-} from '../hooks';
-import { FILTERS_CONFIG } from './configs';
-import { filterAndSort } from './filterAndSort';
+} from '../../../hooks';
+import { FILTERS_CONFIG } from '../../configs';
+import { filterAndSort } from '../../filterAndSort';
 
 const hydrateLocations = (
   locations,
